+++ e2e/tests/plugins.e2e.js
@@ -9,7 +9,6 @@ import {
   GIPHY_PLUGIN,
 } from '../cypress/dataHooks';
 import { DEFAULT_DESKTOP_BROWSERS } from '../tests/constants';
-import linkPreviewMockRes from './linkPreviewMockRes.json';
 
 const eyesOpen = ({
   test: {
@@ -460,28 +459,33 @@ describe('plugins', () => {
     });
     after(() => cy.eyesClose());
 
-    beforeEach('load editor', () => cy.loadEditorAndViewer('linkPreview'));
+    beforeEach('load editor', () => cy.loadEditorAndViewer('link-preview'));
 
     it('change link preview settings', function() {
       cy.openPluginToolbar(PLUGIN_COMPONENT.LINK_PREVIEW);
       cy.setLinkSettings();
-      cy.focusEditor();
+      cy.triggerLinkPreviewViewerUpdate();
+      cy.eyesCheckWindow(this.test.title);
     });
     it('convert link preview to regular link', function() {
       cy.openPluginToolbar(PLUGIN_COMPONENT.LINK_PREVIEW);
-      cy.get(`[data-hook=baseToolbarButton_replaceToLink][tabindex!=-1]`).click();
-      cy.moveCursorToEnd();
-      cy.focusEditor();
+      cy.clickToolbarButton('baseToolbarButton_replaceToLink');
+      cy.triggerLinkPreviewViewerUpdate();
+      cy.eyesCheckWindow(this.test.title);
     });
     it('backspace key should convert link preview to regular link', function() {
-      cy.moveCursorToEnd().type('{backspace}');
-      cy.focusEditor();
+      cy.focusEditor()
+        .type('{downarrow}{downarrow}')
+        .type('{backspace}');
+      cy.triggerLinkPreviewViewerUpdate();
+      cy.eyesCheckWindow(this.test.title);
     });
     it('delete link preview', function() {
       cy.moveCursorToStart();
-      cy.openPluginToolbar(PLUGIN_COMPONENT.LINK_PREVIEW);
-      cy.get(`[data-hook=blockButton_delete][tabindex!=-1]`).click();
-      cy.moveCursorToEnd();
+      cy.openPluginToolbar(PLUGIN_COMPONENT.LINK_PREVIEW).wait(100);
+      cy.clickToolbarButton('blockButton_delete');
+      cy.triggerLinkPreviewViewerUpdate();
+      cy.eyesCheckWindow(this.test.title);
     });
   });
 
@@ -494,10 +498,12 @@ describe('plugins', () => {
 
     it('should create link preview from link after enter key', function() {
       cy.insertLinkAndEnter('www.wix.com');
+      cy.eyesCheckWindow(this.test.title);
     });
 
     it('should embed link that supports embed', function() {
-      cy.insertLinkAndEnter('www.instagram.com');
+      cy.insertLinkAndEnter('www.mockUrl.com');
+      cy.eyesCheckWindow(this.test.title);
     });
   });
 });