+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 ### :bug: Bug Fix
 - `gallery`
   - [#877](https://github.com/wix-incubator/rich-content/pull/877) fix gallery plugin blurry pictures
+- `link-preview`
+  - [#871](https://github.com/wix-incubator/rich-content/pull/871) embed link width 
 
 ### :house: Internal
 - `general`